Fix bug in checking for consistent module dependencies
Fixed a bug in the calculation of having a consistent module schedule when comparing ordering of modules on Paths with dependencies of modules based on consumes.
The old code did not handle consumesMany and sometimes missed inconsistencies when multiple Paths were involved.
The new code no longer uses the old module modulesDependentUpon method and instead uses the more accurate modulesWhoseProductsAreConsumed.
