+++ FWCore/Framework/interface/EDConsumerBase.h
@@ -84,11 +84,6 @@ namespace edm {
     typedef ProductLabels Labels;
     void labelsForToken(EDGetToken iToken, Labels& oLabels) const;
     
-    void modulesDependentUpon(std::string const& iProcessName,
-                              std::string const& iModuleLabel,
-                              bool iPrint,
-                              std::vector<char const*>& oModuleLabels) const;
-
     void modulesWhoseProductsAreConsumed(std::vector<ModuleDescription const*>& modules,
                                          ProductRegistry const& preg,
                                          std::map<std::string, ModuleDescription const*> const& labelsToDesc,
