+++ FWCore/Framework/interface/stream/EDAnalyzerAdaptorBase.h
@@ -93,11 +93,6 @@ namespace edm {
       
       const EDConsumerBase* consumer() const;
       
-      void modulesDependentUpon(std::string const& iProcessName,
-                                std::string const& iModuleLabel,
-                                bool iPrint,
-                                std::vector<char const*>& oModuleLabels) const;
-
       void modulesWhoseProductsAreConsumed(std::vector<ModuleDescription const*>& modules,
                                            ProductRegistry const& preg,
                                            std::map<std::string, ModuleDescription const*> const& labelsToDesc,
