+++ FWCore/Framework/src/EDConsumerBase.cc
@@ -383,35 +383,6 @@ namespace {
   };
 }
 
-void
-EDConsumerBase::modulesDependentUpon(std::string const& iProcessName,
-                                     std::string const& iModuleLabel,
-                                     bool iPrint,
-                                     std::vector<char const*>& oModuleLabels) const {
-  std::set<char const*, CharStarComp> uniqueModules;
-  for(unsigned int index = 0, iEnd = m_tokenInfo.size(); index < iEnd; ++index) {
-    auto const& info = m_tokenInfo.get<kLookupInfo>(index);
-    if(not info.m_index.skipCurrentProcess()) {
-      auto const& labels = m_tokenInfo.get<kLabels>(index);
-      unsigned int const start = labels.m_startOfModuleLabel;
-      char const* processName = &(m_tokenLabels[start+labels.m_deltaToProcessName]);
-      if(iPrint) {
-        LogAbsolute("ModuleDependency") << "ModuleDependency '" << iModuleLabel <<
-        "' may consume product of type '" << info.m_type.className() << 
-        "' with input tag '" << &(m_tokenLabels[start]) <<
-        ':' << &(m_tokenLabels[start+labels.m_deltaToProductInstance]) <<
-        ':' << processName << "'";
-      }
-      if((not processName) or processName[0]==0 or iProcessName == processName) {
-        uniqueModules.insert(&(m_tokenLabels[start]));
-      }
-    }
-  }
-  
-  oModuleLabels = std::vector<const char*>(uniqueModules.begin(),uniqueModules.end());
-}
-
-
 namespace {
   void
   insertFoundModuleLabel(const char* consumedModuleLabel,
