+++ FWCore/Framework/src/stream/EDAnalyzerAdaptorBase.cc
@@ -115,15 +115,6 @@ EDAnalyzerAdaptorBase::consumer() const {
 }
 
 void
-EDAnalyzerAdaptorBase::modulesDependentUpon(std::string const& iProcessName,
-                                            std::string const& iModuleLabel,
-                                            bool iPrint,
-                                            std::vector<char const*>& oModuleLabels) const {
-  assert(not m_streamModules.empty());
-  return m_streamModules[0]->modulesDependentUpon(iProcessName, iModuleLabel, iPrint, oModuleLabels);
-}
-
-void
 EDAnalyzerAdaptorBase::modulesWhoseProductsAreConsumed(std::vector<ModuleDescription const*>& modules,
                                                        ProductRegistry const& preg,
                                                        std::map<std::string, ModuleDescription const*> const& labelsToDesc,
