+++ FWCore/Framework/src/stream/ProducingModuleAdaptorBase.cc
@@ -106,16 +106,6 @@ namespace edm {
 
     template< typename T>
     void
-    ProducingModuleAdaptorBase<T>::modulesDependentUpon(std::string const& iProcessName,
-                                                        std::string const& iModuleLabel,
-                                                        bool iPrint,
-                                                        std::vector<char const*>& oModuleLabels) const {
-      assert(not m_streamModules.empty());
-      return m_streamModules[0]->modulesDependentUpon(iProcessName, iModuleLabel, iPrint, oModuleLabels);
-    }
-
-    template< typename T>
-    void
     ProducingModuleAdaptorBase<T>::modulesWhoseProductsAreConsumed(std::vector<ModuleDescription const*>& modules,
                                                                    ProductRegistry const& preg,
                                                                    std::map<std::string, ModuleDescription const*> const& labelsToDesc,
